% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plotting fatal injuries}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric value defining the number wanted to be analysed.
Possible entries are 1 to 56}

\item{year}{A integer value defining the corresponding year}
}
\value{
Returning a plot which shows the accidents as points in a state map.
If the '\code{state.num}' is invalid an error returns.
If there a no accidents there will be no plot but a message.
}
\description{
Plotting fatal injuries by defing a state number and a year.
\code{\link{make_filename}} and \code{\link{fars_read}} are used to read the files.
}
\examples{
fars_map_state(1, 2013)
fars_map_state(30, 2013)

}
