% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data.R
\name{fars_read}
\alias{fars_read}
\title{Read fatal injuries data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string defing the filename to be loaded}
}
\value{
A tibble object. If the filename does not excist an error returns.
}
\description{
This is a functions which reads a CSV file with data
from the US National Highway Traffic Safety Administration's.
The file needs to be located in your working directory.
}
\examples{
fars_read("accident_2013.csv.bz2")
fars_read("accident_2014.csv.bz2")

}
