% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reading multiple fars files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integer values defining years to be loaded}
}
\value{
Returning lists of dataframes - one per defined year -
with the columns MONTH and year. If the year is invalid an error returns.
}
\description{
Returning the column MONTH and year for given years.
\code{\link{make_filename}} and \code{\link{fars_read}} are used to read the files
}
\examples{
fars_read_years(c(2013, 2015))
fars_read_years(2013:2015)

}
