% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Read and summarize all fatal injuries by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of integer values defining years to be loaded}
}
\value{
Returning a dataframe with the summarized injuries.
Month will be shown as rows and years as columns
}
\description{
Summarizing the fatal injuries per month and year by given years.
Therefore \code{\link{fars_read_years}} is used.
To exceute the functions the following functions need to be imported:
}
\examples{
fars_summarize_years(c(2013, 2015))
fars_summarize_years(c(2014, 2015))

}
