% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data.R
\name{make_filename}
\alias{make_filename}
\title{Creating a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A integer value defining the corresponding year}
}
\value{
Returning a string with the full filename
}
\description{
This function creates a filename by a given year.
Therefore the selected year is embedded in a standardised file name format
}
\examples{
make_filename(2020)

}
